refactor driver code into its own file
